requirements.txt
fastapi
uvicorn
google-ads
